:bug: Fix mistakes in the query
