+++ app/models.py
@@ -2,8 +2,6 @@ from pydantic import BaseModel,UrlStr,Schema, PositiveInt
 from datetime import datetime
 from typing import List
 from enum import Enum
-from sqlalchemy import ForeignKey
-from sqlalchemy.orm import relationship
 from sqlalchemy import Boolean, Column, Integer, String, create_engine
 
 
@@ -13,9 +11,11 @@ class TypeCalque(Enum):
 
 class Calque(BaseModel):
     id:int = Schema (..., gt=0,description="Id de l'oeuvre")
-    typeCalque: TypeCalque = Schema (..., description="Type de calque (composition, anecdote ...")
+    typeCalque:TypeCalque = Schema (..., description="Type de calque (composition, anecdote ...")
     description:str =Schema (...,min_length=1, description="Description du calque")
-    urlCalque: UrlStr = Schema (..., description="Url de l'image du calque")
+    urlCalque:UrlStr = Schema (..., description="Url de l'image du calque")
+    urlAudio:UrlStr = Schema (..., description="Url de l'audio du calque")
+    oeuvre_id:int = Schema(..., gt=0, description="Id de l'oeuvre" )
 
 class Oeuvre(BaseModel):
     id:int = Schema(..., gt=0, description="Id de l'oeuvre")
@@ -26,27 +26,6 @@ class Oeuvre(BaseModel):
     largeur:PositiveInt = Schema (..., description="Largeur de l'oeuvre en cm")
     annee: int = Schema(..., description="Année de réalisation")
     urlCible:UrlStr = Schema (...,min_length=1, description="Url de l'image du tableau")
+    urlAudio: UrlStr = Schema (..., description="Url de l'audio du tableau")
     calques:List[Calque] = Schema (..., description="Calques contenant des informations sur l'oeuvre")
-
-class OeuvreDb(Base):
-    __tablename__ = "oeuvres"
-    id = Column(Integer, primary_key=True, nullable=False)
-    titre = Column(Integer, primary_key=True, nullable=False)
-    auteur = Column(String, nullable=False)
-    technique = Column(String, nullable=False)
-    hauteur = Column(Integer, nullable = False)
-    largeur = Column(Integer, nullable = False)
-    annee = Column(Integer, nullable = False)
-    urlCible = Column(String, nullable=False)
-    calques_id = Column(String, ForeignKey("calques.id"), nullable=False)
-    calques = relationship("Calque", back_populates="oeuvre")
-
-class CalqueDb(Base):
-    __tablename__ = "calques"
-    id = Column(Integer, primary_key=True, nullable=False)
-    typeCalque = Column(String,nullable=False)
-    description = Column(String, nullable=False)
-    urlCalque = Column(String, nullable=False)
-    oeuvres_id = Column(String, ForeignKey("oeuvres.id"), nullable=False)
-    oeuvres = relationship("Oeuvre", back_populates="calque") #mettre un DB après Oeuvre ?
